09.08

+++ src/main/java/ru/itis/springbootsimbirsoft/service/MessageServiceImpl.java
@@ -4,6 +4,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
 import ru.itis.springbootsimbirsoft.domain.entity.Messages;
+import ru.itis.springbootsimbirsoft.domain.enums.StateActive;
+import ru.itis.springbootsimbirsoft.domain.enums.StateRole;
+import ru.itis.springbootsimbirsoft.repository.AccountRepository;
 import ru.itis.springbootsimbirsoft.repository.MessageRepository;
 
 import java.util.Date;
@@ -11,6 +14,10 @@ import java.util.List;
 
 @Component
 public class MessageServiceImpl implements MessageService {
+
+    @Autowired
+    private AccountRepository accountRepository;
+
     @Autowired
     private MessageRepository messageRepository;
 
@@ -23,16 +30,18 @@ public class MessageServiceImpl implements MessageService {
     }
 
     @Override
-    public void writeMessage(Messages form) {
-        Date date = new Date();
-        Messages newMessage = Messages.builder()
-                .account(form.getAccount())
-                .date(date)
-                .room(form.getRoom())
-                .text(form.getText())
-                .build();
+    public void writeMessage(Messages form, Long author) {
+        if (accountRepository.findFirstById(author).getStateActive() == StateActive.ACTIVE) {
+            Date date = new Date();
+            Messages newMessage = Messages.builder()
+                    .account(form.getAccount())
+                    .date(date)
+                    .room(form.getRoom())
+                    .text(form.getText())
+                    .build();
 
-        messageRepository.save(newMessage);
+            messageRepository.save(newMessage);
+        }
     }
 
     @Override
@@ -54,7 +63,10 @@ public class MessageServiceImpl implements MessageService {
     }
 
     @Override
-    public void deleteMessage (String text){
-        messageRepository.delete(messageRepository.findFirstByText(text));
+    public void deleteMessage (String text, Long adminId) {
+        if ((accountRepository.findFirstById(adminId).getRole() == StateRole.ADMIN) ||
+                (accountRepository.findFirstById(adminId).getRole() == StateRole.MODERATOR)) {
+            messageRepository.delete(messageRepository.findFirstByText(text));
+        }
     }
 }
